New ideation form: add main + take out go back button from main
